+++ packages/frontend/src/components/praise/Praise.tsx
@@ -3,10 +3,10 @@ import { Tooltip } from '@mui/material';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faStar } from '@fortawesome/free-solid-svg-icons';
 import { useRecoilValue } from 'recoil';
-import { getMarkdownText } from '@/components/MarkdownText';
+import { getMarkdownText } from '@/utils/parser';
 import { ForwarderTooltip } from '@/components/praise/ForwarderTooltip';
 import { UserAvatar } from '@/components/user/UserAvatar';
-import { InlineLabel } from '@/components/InlineLabel';
+import { InlineLabel } from '@/components/ui/InlineLabel';
 import { classNames } from '@/utils/index';
 import {
   localizeAndFormatIsoDate,
@@ -20,7 +20,7 @@ import { UserName } from '@/components/user/UserName';
 import { useQuantifyPraise } from '@/model/praise';
 import { SourceName } from './SourceName';
 import { UserAvatarAndName } from '../user/UserAvatarAndName';
-import { InlineLabelClosable } from '../InlineLabelClosable';
+import { InlineLabelClosable } from '../ui/InlineLabelClosable';
 
 interface Props {
   praise: PraiseDto;
@@ -145,7 +145,8 @@ export const Praise = ({
               }}
               className={classNames(
                 dismissed ? 'line-through' : '',
-                shortDuplicatePraiseId ? 'text-warm-gray-400' : ''
+                shortDuplicatePraiseId ? 'text-warm-gray-400' : '',
+                'break-words'
               )}
             ></span>
           </div>