+++ packages/frontend/src/layouts/SignMessageLayout.tsx
@@ -4,7 +4,7 @@ import { ConnectButton } from '@rainbow-me/rainbowkit';
 import { useAccount } from 'wagmi';
 import { toast } from 'react-hot-toast';
 import { SignMessageButton } from '@/components/auth/SignMessageButton';
-import { LoaderSpinner } from '@/components/LoaderSpinner';
+import { LoaderSpinner } from '@/components/ui/LoaderSpinner';
 import { EthAccount } from '@/components/account/EthAccount';
 
 interface Props {
@@ -28,7 +28,7 @@ export const SignMessageLayout = ({
         <div className="w-full p-5 text-2xl font-bold">
           <FontAwesomeIcon icon={faPrayingHands} size="1x" className="m-2" />
         </div>
-        <div className="p-4 py-8 m-auto space-y-8 border border-solid rounded-lg shadow-sm bg-warm-gray-50 w-96 dark:bg-slate-900">
+        <div className="p-8 py-8 m-auto space-y-8 border-none rounded-none shadow-none md:p-4 md:border md:border-solid md:rounded-lg md:shadow-sm bg-none md:bg-warm-gray-50 w-96 dark:bg-slate-900">
           {children}
 
           {isLoading && !data && !message ? (