+++ packages/frontend/src/pages/Periods/PeriodsPage.tsx
@@ -4,9 +4,12 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { ActiveNoticesBoard } from '@/components/periods/ActiveNoticesBoard';
-import { BreadCrumb } from '@/components/BreadCrumb';
+import { BreadCrumb } from '@/components/ui/BreadCrumb';
 import { AdminOnly } from '@/components/auth/AdminOnly';
 import { AllPeriods } from '@/model/periods';
+import { Page } from '@/components/ui/Page';
+import { Button } from '@/components/ui/Button';
+import { Box } from '@/components/ui/Box';
 import { PeriodsTable } from './components/Table';
 
 const PeriodsPage = (): JSX.Element | null => {
@@ -15,27 +18,27 @@ const PeriodsPage = (): JSX.Element | null => {
   if (!allPeriods) return null;
 
   return (
-    <div className="praise-page">
+    <Page>
       <BreadCrumb name="Quantification periods" icon={faCalendarAlt} />
 
       <ActiveNoticesBoard />
 
-      <div className="px-0 praise-box">
+      <Box className="px-0">
         <AdminOnly>
           <div className="px-5 mb-2 text-right">
             <Link to="/periods/createupdate">
-              <button className="praise-button" id="create-period-button">
+              <Button id="create-period-button">
                 <FontAwesomeIcon icon={faPlus} size="1x" className="mr-2" />
                 Create period
-              </button>
+              </Button>
             </Link>
           </div>
         </AdminOnly>
         <React.Suspense fallback={null}>
           <PeriodsTable />
         </React.Suspense>
-      </div>
-    </div>
+      </Box>
+    </Page>
   );
 };