+++ packages/frontend/src/pages/Users/components/UsersStatistics.tsx
@@ -1,15 +1,17 @@
-import { useState } from 'react';
-import { useRecoilValue } from 'recoil';
-import { CSVLink } from 'react-csv';
 import { faDownload } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import type { UserDto } from 'api/dist/user/types';
+import { useState } from 'react';
+import { CSVLink } from 'react-csv';
+import { useRecoilValue } from 'recoil';
+
 import {
   AllAdminUsers,
   AllForwarderUsers,
   AllQuantifierUsers,
   AllUsers,
 } from '@/model/users';
+import { Button } from '@/components/ui/Button';
 
 interface DownloadUserDto extends Omit<UserDto, 'accounts'> {
   discordName?: string;
@@ -62,7 +64,7 @@ export const UsersStatistics = (): JSX.Element => {
       <div>Quantifiers: {allQuantifierUsers?.length}</div>
       <div className="flex w-full mt-5">
         {allUsers && (
-          <div className="praise-button">
+          <Button>
             <CSVLink
               className="no-underline"
               data={downloadData}
@@ -73,7 +75,7 @@ export const UsersStatistics = (): JSX.Element => {
               <FontAwesomeIcon icon={faDownload} size="1x" className="mr-2" />
               Export
             </CSVLink>
-          </div>
+          </Button>
         )}
       </div>
     </>