+++ setup.sh
@@ -1,2 +1,9 @@
-docker build -t praise-setup -f ./packages/setup/Dockerfile .
+#!/bin/bash
+IMAGE=$( docker images | grep praise-setup )
+if [ -z "$IMAGE" ]
+then
+  echo "Building setup image..."
+  docker build -t praise-setup -q -f ./packages/setup/Dockerfile .
+fi
+
 docker run -it -v $(pwd):/usr/praise praise-setup